pkg/utils: Temporal removal of tag parsing in the Fedora images
Parsing of the tag in the Fedora images has to be temporarily
removed to use images with tags such as `42-aarch62`, '42-ppc64Ie',
etc. This came up from the demand for Toolbx to support images
with different architectures that the host has.

https://github.com/containers/toolbox/issues/1654

Such as this is a workaround, in the future, it will be necessary
to extend the parsing to support different tags other than
numerical.

Signed-off-by: Dalibor Kricka <dalidalk@seznam.cz>